refactor: move services directory from core to app directory
