+++ src/app/pages/catalog/catalog.component.spec.ts
@@ -1,11 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { BehaviorSubject, of } from 'rxjs';
 import { CatalogComponent } from './catalog.component';
-import { CategoriesService } from '../../core/services/categories.service';
-import { CategoriesStoreService } from '../../core/services/categories-store.service';
-import { ErrorHandlerService } from '../../core/services/error-handler.service';
 import { provideRouter } from '@angular/router';
-import { RandomizationService } from '../../core/services/randomization.service';
 import { UiButtonComponent } from '../../shared/ui-kit/ui-button/ui-button.component';
 import { UiQuizCardComponent } from '../../shared/ui-kit/ui-quiz-card/ui-quiz-card.component';
 import { UiSpinnerComponent } from '../../shared/ui-kit/ui-spinner/ui-spinner.component';
@@ -13,6 +9,10 @@ import { UiErrorNotificationComponent } from '../../shared/ui-kit/ui-error-notif
 import { By } from '@angular/platform-browser';
 import { provideAngularSvgIcon } from 'angular-svg-icon';
 import { provideHttpClient } from '@angular/common/http';
+import { CategoriesService } from '../../services/categories-service/categories.service';
+import { CategoriesStoreService } from '../../services/categories-store-service/categories-store.service';
+import { ErrorHandlerService } from '../../services/error-handler-service/error-handler.service';
+import { RandomizationService } from '../../services/randomization-service/randomization.service';
 
 describe('CatalogComponent Template', () => {
   let component: CatalogComponent;
