+++ src/app/pages/statistics/statistics.component.ts
@@ -1,7 +1,6 @@
 import { Component, inject, OnInit, signal } from '@angular/core';
 import { UiButtonComponent } from '../../shared/ui-kit/ui-button/ui-button.component';
 import {} from '../../shared/ui-kit/ui-chart-legend-item/ui-chart-legend-item.component';
-import { StatisticsService } from '../../core/services/statistics.service';
 import { INITIAL_QUIZ_STATISTIC } from '../../shared/models/quiz-statistic.model';
 import {
   INITIAL_USER_STATISTIC,
@@ -11,6 +10,7 @@ import { LastQuizData } from '../../shared/models/last-quiz-data.model';
 import { QuizFeedbackComponent } from '../../shared/ui-kit/ui-quiz-feedback/ui-quiz-feedback.component';
 import { TimeFormatPipe } from '../../shared/ui-kit/ui-quiz-feedback/time-format.pipe';
 import { UiDonutChartComponent } from '../../shared/ui-kit/ui-donut-chart/ui-donut-chart.component';
+import { StatisticsService } from '../../services/statistics-service/statistics.service';
 
 @Component({
   selector: 'quiz-statistics',
