+++ src/app/services/questions-service/questions.service.ts
@@ -1,7 +1,7 @@
 import { inject, Injectable } from '@angular/core';
 import { Observable, catchError, map } from 'rxjs';
-import { ApiService } from './api.service';
-import { ErrorHandlerService } from './error-handler.service';
+import { ApiService } from '../api-service/api.service';
+import { ErrorHandlerService } from '../error-handler-service/error-handler.service';
 import { QuestionApiResponse } from '../../shared/models/question-api-response.model';
 import { Question } from '../../shared/models/question.model';
