+++ src/app/services/randomization-service/randomization.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
-import { RandomizationService } from './randomization.service';
 import { QuizCardColors } from '../../shared/enums/quiz-card-colors.enums';
+import { RandomizationService } from './randomization.service';
 
 describe('RandomizationService', () => {
   let service: RandomizationService;
